% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jfids.R
\name{gmr_forfile}
\alias{gmr_forfile}
\title{Return the GMR code (ie plate row column) for a file}
\usage{
gmr_forfile(f)
}
\arguments{
\item{f}{one or more paths}
}
\value{
character vector GMR code
}
\description{
Return the GMR code (ie plate row column) for a file
}
\seealso{
\code{\link{jfimagestem_forfile}}
}

