% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jfids.R
\name{jfimagestem_forfile}
\alias{jfimagestem_forfile}
\title{Return the janelia image name for a path (excluding any file extension)}
\usage{
jfimagestem_forfile(file, checkExists = FALSE)
}
\arguments{
\item{file}{Path to file(s) to test}

\item{checkExists}{Check that calculated image stem exists (TODO)}
}
\value{
Character vector of image names (or NA if checkExists fails)
}
\description{
Return the janelia image name for a path (excluding any file extension)
}
\examples{
stopifnot(jfimagestem_forfile("Registration.GHT/affine/JFRC2_GMR_10A06_AE_01_08-fA01b_C100709_20100709225242046_01_ght.list")=="GMR_10A06_AE_01_08-fA01b_C100709_20100709225242046")
}
\seealso{
\code{\link{jfshortname_forfile}}
}

